save gas

+++ packages/sharing-smart-contract/contracts/modules/Subscription.sol
@@ -29,14 +29,14 @@ contract Subscription is Collection {
 
     struct SubscriptionParams {
         uint112 price; // 112 bit allows for 10^15 eth
-        uint48 duration; // 48 bit allows 89194 years of delay
+        uint48 duration; // 48 bit allows (2**48)/60/60/24/365 of delay
     }
 
     /***************************************************************************
      *                        event/modifier                                   *
      ***************************************************************************/
     event NewSubscriptionParams(SubscriptionParams subscriptionParams);
-    event NewSubscription(address indexed subscriber, uint256 endDate);
+    event NewSubscription(address indexed subscriber, uint48 endDate);
     event AddProtectedDataForSubscription(uint256 _collectionId, address _protectedData);
 
     /***************************************************************************
